// frontend/src/features/site/types.ts
export interface SiteSetting {
  id: number
  siteName: string
  siteSubtitle: string | null
  siteDescription: string | null
  siteKeywords: string | null
  logo: string | null
  favicon: string | null
  githubUrl: string | null
  aboutMarkdown: string | null
  icp: string | null
  publicEmail: string | null
}